Temperature Median
**User Story #1**
As an anonymous end user, I want to be able to enter any number of
temperture records by entering a value and then clicking on the "Add"
button.

**User Story #2**
As an anonymous end user,  When I click on the "Get Media Temperture"
button, I want to be able to see the median of the entered tempertures
records that I have entered.

**User Story #3**
As an anonymous end user, I want to be prevented from entering invalid
value such as a strings in the "temprature" field.
The UI should inform me that I tried an invalid action.

##Unit Test Sample Data
`[5, 1, -7, 7, 8, 3] --> [-7, 1, 3, 5, 7, 8] = (3+5)/2 = 4`
`[5, 1, -7, 7, 8, 3, 9] --> [-7, 1, 3, 5, 7, 8, 9] = 5`